requirements.txt
onnxruntime
numpy
pandas
flask
pillow
ollama
huggingface-hub
python-dotenv